Pause button, edited start screen, score&lives

+++ Raindrops_Game/Raindrops_Game.pde
@@ -12,13 +12,21 @@ PImage storm;
 //boolean to start and pause game
 boolean run;
 //variable that, once increased to a certain amount, causes you to lose the game
-int life;
+int life = 1;
 //variable controlling the fill of Game Over text
 float x;
+//to pause the game
+boolean pause;
+//gameover
+boolean gameOver;
+//lives
+int lives = 5;
 
 void setup() {
   //boolean starts as true
   run = false;
+  pause = false;
+  gameOver = false;
   //storm background
   storm = loadImage("Rainstorm background.png");
   //change color mode so that the fill will be different colors
@@ -37,23 +45,32 @@ void draw() {
     background(storm);
     C.display();
     //text size for the score
-    textSize(50);
-    text(C.score, 20, 100);
+    textSize(20);
+    text("Score:", 40, 375);
+    text(C.score, 95, 375);
+    text("Lives:", 500, 375);
+    text(lives, 550, 375);
     for (int i = 0; i < index; i++) {
       //to display the raindrops
       Drops[i].display();
       //to ensure the raindrops drop
-      Drops[i].drop();
-      //once the raindrops come in contact with the catcher they are seny back to reset
-      C.catchDrop(Drops[i]);
+      if (!pause) {
+        Drops[i].drop();
+        C.catchDrop(Drops[i]);
+      }//once the raindrops come in contact with the catcher they are sent back to reset
+
       if (Drops[i].loc.y >= height) {
         life++;
+        lives--;
       }
       //Reset, as stated before
       Drops[i].reset();
       //this if statement describes that, once you lose a certain amount of lives, it's Game Over, adn the Game over end screen is displayed
       if (life > 5) {
-        //Black Background
+        gameOver = true;
+      }
+      //Black Background
+      if (gameOver) {
         background(0);
         //Center text
         textAlign(CENTER);
@@ -72,6 +89,7 @@ void draw() {
         text(C.score, width/2, 2*height/3);
       }
     }
+
     //Timer Class properties 
     if (A.time() >= 1000 && index < Drops.length) {  
       index++;
@@ -79,17 +97,22 @@ void draw() {
     }
   }
   else {
-    background (255);
-    text("press any key to start", width/2, height/2);
+    background (255, 0, 0);
+    textSize(30);
+    textAlign(CENTER);
+    text("Press any key to START", width/2, height/2);
   }
-}
-//when the key is pressed the 
-void mousePressed() {
-  if (life <= 5) {    
+
+  if (pause && run) {    
     textAlign(CENTER);
     textSize(45);
     text("Paused", width/2, height/2);
   }
+}
+
+//when the key is pressed the 
+void mousePressed() {
+  pause = !pause;
 }
 
 void keyPressed() {
